Avoid rendering fills if empty elements inside
