+++ packages/editor/src/components/reusable-blocks-buttons/reusable-block-delete-button.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import { noop } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import { compose } from '@wordpress/compose';
@@ -11,6 +6,7 @@ import { MenuItem } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { isReusableBlock } from '@wordpress/blocks';
 import { withSelect, withDispatch } from '@wordpress/data';
+import { BlockSettingsMenuControls } from '@wordpress/block-editor';
 
 export function ReusableBlockDeleteButton( {
 	isVisible,
@@ -22,9 +18,21 @@ export function ReusableBlockDeleteButton( {
 	}
 
 	return (
-		<MenuItem disabled={ isDisabled } onClick={ () => onDelete() }>
-			{ __( 'Remove from Reusable blocks' ) }
-		</MenuItem>
+		<BlockSettingsMenuControls>
+			{ ( { onClose } ) => (
+				<MenuItem
+					disabled={ isDisabled }
+					onClick={ () => {
+						const hasConfirmed = onDelete();
+						if ( hasConfirmed ) {
+							onClose();
+						}
+					} }
+				>
+					{ __( 'Remove from Reusable blocks' ) }
+				</MenuItem>
+			) }
+		</BlockSettingsMenuControls>
 	);
 }
 
@@ -49,7 +57,7 @@ export default compose( [
 			isDisabled: reusableBlock && reusableBlock.isTemporary,
 		};
 	} ),
-	withDispatch( ( dispatch, { clientId, onToggle = noop }, { select } ) => {
+	withDispatch( ( dispatch, { clientId }, { select } ) => {
 		const {
 			__experimentalDeleteReusableBlock: deleteReusableBlock,
 		} = dispatch( 'core/editor' );
@@ -69,8 +77,9 @@ export default compose( [
 				if ( hasConfirmed ) {
 					const block = getBlock( clientId );
 					deleteReusableBlock( block.attributes.ref );
-					onToggle();
 				}
+
+				return hasConfirmed;
 			},
 		};
 	} ),
